computer_vision/requirements.txt
opencv-python==4.9.0.80
numpy==1.23.4
Pillow==9.2.0
pyserial
